Login/requirement.txt
aiofiles==0.5.0
aiosmtplib==1.1.4
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cryptography==3.2.1
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.1
fastapi==0.61.1
fastapi-mail==0.3.0.1
gitdb==4.0.5
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.3
orjson==3.4.0
pandas==1.1.4
passlib==1.7.4
promise==2.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
Rx==1.6.1
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.20
starlette==0.13.6
ujson==3.2.0
urllib3==1.25.10
uvicorn==0.11.8
websockets==8.1
